Check if PEM bundle is valid after downloading
